# backend/src/main/java/com/examsystem/modules/auth/dto/ChangePasswordRequest.java
package com.examsystem.modules.auth.dto;

import jakarta.validation.constraints.NotBlank;

public record ChangePasswordRequest(
        @NotBlank String currentPassword,
        @NotBlank String newPassword
) {
}