# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 企业内部练习与考试系统 API
  version: 1.0.0
  description: |
    MVP 完整接口规格。全局约定见 [00-API设计说明.md](./00-API设计说明.md)。
    分域行为见 01~07 Markdown 文档。
  contact:
    name: Exam System Team
servers:
  - url: /api/v1
    description: 单租户企业内网 API
...
security:
  - sessionAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: authLogin
      summary: 工号密码登录
      security: []
      x-fr-ids: [FR-AUTH-01]
      x-page-ids: [MP-01, EX-01, AD-01]
      x-sys-ids: [SYS-01]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /auth/session:
    get:
      tags: [Auth]
      operationId: authGetSession
      summary: 查询当前会话
      x-fr-ids: [FR-AUTH-01, FR-AUTH-06]
      responses:
        '200':
          description: 会话有效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/change-password:
    post:
      tags: [Auth]
      operationId: authChangePassword
      summary: 修改密码
      x-fr-ids: [FR-AUTH-02]
      x-page-ids: [MP-01, EX-01, AD-01, MP-12]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
  /exams/{id}/attempts:
    post:
      tags: [Attempt]
      operationId: startExamAttempt
      summary: 幂等开卷
      x-fr-ids: [FR-ATT-01, FR-ATT-02, FR-ATT-03]
      x-page-ids: [EX-03, EX-04]
      x-sys-ids: [SYS-03]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 新开或返回原尝试
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExamAttemptStartResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /attempts/{id}:
    get:
      tags: [Attempt]
      operationId: getExamAttempt
      summary: 尝试详情（恢复+计时）
      x-fr-ids: [FR-ANS-02]
      x-page-ids: [EX-04]
      x-sys-ids: [SYS-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 含 timing 与已确认答案
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExamAttemptDetail'
  /attempts/{id}/paper:
    get:
      tags: [Attempt]
      operationId: getExamPaper
      summary: 固定试卷（无答案）
      x-fr-ids: [FR-ATT-02]
      x-page-ids: [EX-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 固定试卷（无标准答案）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Paper'
  /attempts/{id}/answers/{itemId}:
    put:
      tags: [Answer]
      operationId: saveExamAnswer
      summary: 版本化保存答案
      x-fr-ids: [FR-ANS-01]
      x-page-ids: [EX-04]
      x-sys-ids: [SYS-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAnswerRequest'
      responses:
        '200':
          description: 答案已确认
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SaveAnswerResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /attempts/{id}/submit:
    post:
      tags: [Answer]
      operationId: submitExamAttempt
      summary: 幂等交卷
      x-fr-ids: [FR-ANS-03, FR-ANS-04]
      x-page-ids: [EX-04]
      x-sys-ids: [SYS-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  enum: [manual, timeout]
                  default: manual
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
        '409':
          description: 未确认答案或冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /attempts/{id}/result:
    get:
      tags: [Score]
      operationId: getAttemptResult
      summary: 受控结果/复盘
      x-fr-ids: [FR-SCR-04]
      x-page-ids: [EX-05, MP-11]
      x-sys-ids: [SYS-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 受控结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EmployeeAttemptResult'
...
  /admin/exams/{id}/attempts/{attemptId}:
    get:
      tags: [Report]
      operationId: getAdminExamAttempt
      summary: 尝试详情（管理端完整）
      x-fr-ids: [FR-REP-02, FR-SCR-03]
      x-page-ids: [AD-15]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
...
  /admin/audit-logs:
    get:
      tags: [Audit]
      operationId: listAuditLogs
      summary: 审计日志查询
      x-fr-ids: [FR-OPS-03]
      x-page-ids: [AD-16]
      x-sys-ids: [SYS-07]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: actorId
          in: query
          schema:
            type: string
        - name: actionType
          in: query
          schema:
            type: string
        - name: targetType
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 亦可用 HttpOnly Session Cookie，二者等价
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 幂等键，见 00-API设计说明
  responses:
    SuccessEmpty:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
    Unauthorized:
      description: 未登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 无权限或账号受限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Conflict:
      description: 版本或状态冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unprocessable:
      description: 业务规则拒绝
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ApiEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      required: [serverNow, requestId]
      properties:
        serverNow:
          type: string
          format: date-time
        requestId:
          type: string
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorCode:
      type: string
      enum:
        - AUTH_INVALID_CREDENTIALS
        - AUTH_ACCOUNT_LOCKED
        - AUTH_ACCOUNT_DISABLED
        - AUTH_SESSION_EXPIRED
        - AUTH_PASSWORD_CHANGE_REQUIRED
        - AUTH_SMS_RATE_LIMITED
        - ORG_DEPARTMENT_NAME_DUPLICATE
        - ORG_LAST_OUTAGE_ADMIN
        - IMP_PREVIEW_STALE
        - PRA_SESSION_ALREADY_ACTIVE
        - SIM_ATTEMPT_ALREADY_ACTIVE
        - ATT_NO_REMAINING_OPPORTUNITY
        - ATT_EXAM_PAUSED
        - ANS_VERSION_CONFLICT
        - ANS_UNCONFIRMED_ANSWERS
        - ANS_ATTEMPT_TERMINATED
        - SCR_SUBMIT_IN_PROGRESS
        - OPS_PROPOSAL_STALE
        - OPS_NOT_AUTHORIZED
        - SEC_FORBIDDEN
    ClientType:
      type: string
      enum: [miniProgram, examWeb, adminWeb]
    LoginRequest:
      type: object
      required: [employeeNo, password, clientType]
      properties:
        employeeNo:
          type: string
        password:
          type: string
          format: password
        clientType:
          $ref: '#/components/schemas/ClientType'
    Session:
      type: object
      properties:
        employeeId:
          type: string
        employeeNo:
          type: string
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
        isAdmin:
          type: boolean
        hasOutageDisposition:
          type: boolean
        mustChangePassword:
          type: boolean
        miniProgramBound:
          type: boolean
    SessionResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                session:
                  $ref: '#/components/schemas/Session'
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          format: password
...
    QuestionType:
      type: string
      enum: [singleChoice, multipleChoice, trueFalse]
...
    AttemptTiming:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        remainingSeconds:
          type: integer
        serverNow:
          type: string
          format: date-time
    PaperItem:
      type: object
      properties:
        itemId:
          type: string
        order:
          type: integer
        questionVersionId:
          type: string
        score:
          type: number
        type:
          $ref: '#/components/schemas/QuestionType'
        stem:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              text:
                type: string
    Paper:
      type: object
      properties:
        attemptId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaperItem'
    SaveAnswerRequest:
      type: object
      required: [answer, answerVersion]
      properties:
        answer:
          type: array
          items:
            type: string
        answerVersion:
          type: integer
          minimum: 1
    SaveStatus:
      type: string
      enum: [saving, saved, failed]
    SaveAnswerResponse:
      type: object
      properties:
        itemId:
          type: string
        confirmedVersion:
          type: integer
        saveStatus:
          $ref: '#/components/schemas/SaveStatus'
        confirmedAt:
          type: string
          format: date-time
    ExamLifecycle:
      type: string
      enum: [draft, notStarted, openForAttempt, closing, ended, cancelled]
    ExamRunStatus:
      type: string
      enum: [normal, paused]
    AttemptStatus:
      type: string
      enum: [inProgress, submitting, completed, voided, terminated]
...
    ExamAttemptStartResponse:
      type: object
      properties:
        attemptId:
          type: string
        attemptNumber:
          type: integer
        attemptStatus:
          $ref: '#/components/schemas/AttemptStatus'
        paper:
          $ref: '#/components/schemas/Paper'
        timing:
          $ref: '#/components/schemas/AttemptTiming'
    ConfirmedAnswer:
      type: object
      properties:
        itemId:
          type: string
        answer:
          type: array
          items:
            type: string
        confirmedVersion:
          type: integer
        saveStatus:
          $ref: '#/components/schemas/SaveStatus'
    ExamAttemptDetail:
      type: object
      properties:
        attemptId:
          type: string
        examId:
          type: string
        attemptStatus:
          $ref: '#/components/schemas/AttemptStatus'
        attemptNumber:
          type: integer
        timing:
          $ref: '#/components/schemas/AttemptTiming'
        confirmedAnswers:
          type: array
          items:
            $ref: '#/components/schemas/ConfirmedAnswer'
        runStatus:
          $ref: '#/components/schemas/ExamRunStatus'
        lifecycle:
          $ref: '#/components/schemas/ExamLifecycle'
    ResultVisibility:
      type: object
      properties:
        summaryVisible:
          type: boolean
        passingScoreVisible:
          type: boolean
        passConclusionVisible:
          type: boolean
        perItemReviewAllowed:
          type: boolean
    EmployeeAttemptResult:
      type: object
      properties:
        attemptId:
          type: string
        visibility:
          $ref: '#/components/schemas/ResultVisibility'
        totalScore:
          type: [number, 'null']
        maxScore:
          type: [number, 'null']
        passingScore:
          type: [number, 'null']
        retakeNeutralMessage:
          type: [string, 'null']
        items:
          type: [array, 'null']
          items:
            type: object
            properties:
              itemId:
                type: string
              isCorrect:
                type: boolean
              userAnswer:
                type: array
                items:
                  type: string
              standardAnswer:
                type: array
                items:
                  type: string
              explanation:
                type: string
...